Change discrete diagonal
If we are decoding immobility times, we should assume the probability of staying in a state is roughly the same as staying in a sharp wave ripple. We can use the fact that mean probability of staying in a state follows a geometric distribution and we have 2 ms time bins.